update cat-eval sample
