+++ js/testapps/cat-eval/package.json
@@ -6,6 +6,8 @@
   "scripts": {
     "start": "node lib/index.js",
     "compile": "tsc",
+    "dev": "tsx --watch src/index.ts",
+    "genkit:dev": "cross-env GENKIT_ENV=dev pnpm dev",
     "build": "pnpm build:clean && pnpm compile",
     "build:clean": "rimraf ./lib",
     "build:watch": "tsc --watch"
@@ -30,6 +32,8 @@
     "pdfjs-dist-legacy": "^1.0.1"
   },
   "devDependencies": {
+    "cross-env": "^7.0.3",
+    "tsx": "^4.7.0",
     "rimraf": "^6.0.1",
     "@types/pdf-parse": "^1.1.4",
     "typescript": "^5.3.3"
