+++ js/testapps/cat-eval/src/setup.ts
@@ -18,7 +18,7 @@ import { z } from 'genkit';
 import { indexPdf } from './pdf_rag.js';
 import { indexPdfFirebase } from './pdf_rag_firebase.js';
 
-import { ai } from './index.js';
+import { ai } from './genkit.js';
 
 const catFacts = ['./docs/sfspca-cat-adoption-handbook-2023.pdf'];
